Refactor repo to feature-first layout
