+++ src/modules/payments/controller.js
@@ -1,0 +1,1 @@
+// empty module
