+++ src/modules/rides/controller.js
@@ -1,0 +1,1 @@
+// empty module
