+++ src/modules/rides/routes.js
@@ -1,0 +1,1 @@
+// empty module
